server/requirements.txt
fastapi[standard]>=0.115.0
uvicorn>=0.34.0
websockets>=14.0
anthropic>=0.49.0
google-cloud-texttospeech>=2.16.0
deepgram-sdk>=3.8.0
silero-vad>=5.1
python-dotenv>=1.0.0
